requirements.txt
httpx
beautifulsoup4
ollama
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
rich
